Go-over on existing migrated code, as well as QC base refactor.
