+++ tsdat/qc/handlers/record_quality_results.py
@@ -36,12 +36,14 @@ class RecordQualityResults(QualityHandler):
         meaning: str
         """A string that describes the test applied."""
 
+        # TODO: Seems like a static method here, should refactor into as such.
         @root_validator(pre=True)
         def deprecate_bit_parameter(cls, values: Dict[str, Any]) -> Dict[str, Any]:
             if "bit" in values:
                 logger.warning("The 'bit' argument is deprecated, please remove it.")
             return values
 
+        # TODO: Seems like a static method here, should refactor into as such.
         @validator("assessment", pre=True)
         def to_lower(cls, assessment: Any) -> str:
             if isinstance(assessment, str):
@@ -64,6 +66,7 @@ class RecordQualityResults(QualityHandler):
         )
         return dataset
 
+    # TODO: Seems like a static method here, should refactor into as such.
     def get_next_bit_number(self, dataset: xr.Dataset, variable_name: str) -> int:
         if (qc_var := dataset.get(f"qc_{variable_name}")) is None:
             return 1
